CLI, bugfixes, improved public API.
It might be possible that argparse became a dependency at some point.
O:)

Files are now copied as they should have, and extracting comments and
stuff from .moon files works again. With the -o option, it’s a lot
easier to make tests.

It’s also possible to use lunadoc from scripts instead of its CLI.
That’s not recommended at this point, but it IS possible.

Note: Dysnomia has a DSL for generating XML/HTML. Better use that than
      Lapis in the future.