DocTree workarounds.
Should really rewrite huge parts of DocTree to improve reference
handling and stuff. Parts of the templates would need update as well.

Right now, having classes in a returned array is not pretty. But
Hestia’s goal is to make anything documentable pretty, so… I’ll have to
deal with it at some point. :(